+++ source/renderer/dialog/preferences.js
@@ -1,0 +1,106 @@
+const ZettlrDialog = require('./zettlr-dialog.js')
+const validate = require('../../common/validate.js')
+
+class PreferencesDialog extends ZettlrDialog {
+  constructor () {
+    super()
+    this._dialog = 'preferences'
+    // TODO: Implement the proceed functionality.
+  }
+
+  preInit (data) {
+    for (let l of data.availableDicts) {
+      if (data.selectedDicts.includes(l)) {
+        // Remove already selected dictionaries from the list
+        data.availableDicts.splice(data.availableDicts.indexOf(l), 1)
+      }
+    }
+    // The template expects a simple string
+    data.attachmentExtensions = data.attachmentExtensions.join(', ')
+    return data
+  }
+
+  postAct () {
+    // Activate the form to be submitted
+    let form = this._modal.find('form#dialog')
+    form.on('submit', (e) => {
+      e.preventDefault()
+      // Give the ZettlrBody object the results
+      // Form: dialog type, values, the originally passed object
+      this.proceed(form.serializeArray())
+    })
+  }
+
+  proceed (data) {
+    // First remove potential error-classes
+    this.getModal().find(`input`).removeClass('has-error')
+
+    let cfg = {}
+
+    // Standard preferences
+    cfg['darkTheme'] = (data.find(elem => elem.name === 'darkTheme') !== undefined)
+    cfg['snippets'] = (data.find(elem => elem.name === 'snippets') !== undefined)
+    cfg['muteLines'] = (data.find(elem => elem.name === 'muteLines') !== undefined)
+    cfg['export.stripIDs'] = (data.find(elem => elem.name === 'export.stripIDs') !== undefined)
+    cfg['export.stripTags'] = (data.find(elem => elem.name === 'export.stripTags') !== undefined)
+    cfg['debug'] = (data.find(elem => elem.name === 'debug') !== undefined)
+
+    // Display checkboxes
+    cfg['display.renderCitations'] = (data.find(elem => elem.name === 'display.renderCitations') !== undefined)
+    cfg['display.renderIframes'] = (data.find(elem => elem.name === 'display.renderIframes') !== undefined)
+    cfg['display.renderImages'] = (data.find(elem => elem.name === 'display.renderImages') !== undefined)
+    cfg['display.renderLinks'] = (data.find(elem => elem.name === 'display.renderLinks') !== undefined)
+    cfg['display.renderMath'] = (data.find(elem => elem.name === 'display.renderMath') !== undefined)
+    cfg['display.renderTasks'] = (data.find(elem => elem.name === 'display.renderTasks') !== undefined)
+
+    cfg['editor.autoCloseBrackets'] = (data.find(elem => elem.name === 'editor.autoCloseBrackets') !== undefined)
+    // Extract selected dictionaries
+    cfg['selectedDicts'] = data.filter(elem => elem.name === 'selectedDicts').map(elem => elem.value)
+
+    // Copy over all other field values from the result set.
+    for (let r of data) {
+      // Only non-missing to not overwrite the checkboxes that ARE checked with a "yes"
+      if (!cfg.hasOwnProperty(r.name)) {
+        // Convert numbers to prevent validation errors.
+        if (!isNaN(r.value) && r.value !== '') r.value = Number(r.value)
+        cfg[r.name] = r.value
+      }
+    }
+
+    // Now finally the attachment extensions.
+    if (cfg.hasOwnProperty('attachmentExtensions')) {
+      let attachments = cfg['attachmentExtensions'].split(',')
+      for (let i = 0; i < attachments.length; i++) {
+        attachments[i] = attachments[i].trim().replace(/[\s]/g, '')
+        if (attachments[i].length < 2) {
+          attachments.splice(i, 1)
+          i--
+          continue
+        }
+        if (attachments[i].charAt(0) !== '.') {
+          attachments[i] = '.' + attachments[i]
+        }
+      }
+      cfg['attachmentExtensions'] = attachments
+    }
+
+    // Validate dat shit.
+    let unvalidated = validate(cfg)
+
+    if (unvalidated.length > 0) {
+      // For brevity reasons only show one at a time (they have to be resolved either way)
+      this.getModal().find('.error-info').text(unvalidated[0].reason)
+      for (let prop of unvalidated) {
+        // Indicate which ones were wrong.
+        this.getModal().find(`input[name="${prop.key}"]`).first().addClass('has-error')
+      }
+      return // Don't try to update falsy settings.
+    }
+
+    // Finally send and close this dialog.
+    global.ipc.send('update-config', cfg)
+    this.close()
+  }
+}
+
+module.exports = PreferencesDialog